Use custom promise implementation to sign URLs
For some reason, the getSignedUrlPromise API doesn't work and throws
the exception: "getSignedUrlPromise is not a function".

This might have to do with an SDK version mismatch, but I'm not sure,
since it's not documented anywhere in the SDK docs.
But this issue mentions something about a V3 release being required:

https://github.com/aws/aws-sdk-js/issues/1008